+++ CHANGELOG.md
@@ -6,10 +6,11 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [vx.x.x] - xxxx-xx-xx
 
+### Added
+
 - feature/1199692948932202 - Add getServer method to request class and interface
 - feature/1199692948932198 - Add getFiles to RequestInterface
 
-### Added
-
 ### Changed
 
+- fixes/1200549021075051 - Fix to comply with typed-collection